Move SendContext to its own component
