+++ llarp/service/endpoint.hpp
@@ -9,6 +9,8 @@
 #include <service/address.hpp>
 #include <service/Identity.hpp>
 #include <service/handler.hpp>
+#include <service/pendingbuffer.hpp>
+#include <service/sendcontext.hpp>
 #include <service/protocol.hpp>
 
 // minimum time between introset shifts
@@ -188,24 +190,6 @@ namespace llarp
       void
       FlushSNodeTraffic();
 
-      struct PendingBuffer
-      {
-        std::vector< byte_t > payload;
-        ProtocolType protocol;
-
-        PendingBuffer(const llarp_buffer_t& buf, ProtocolType t)
-            : payload(buf.sz), protocol(t)
-        {
-          memcpy(payload.data(), buf.base, buf.sz);
-        }
-
-        ManagedBuffer
-        Buffer()
-        {
-          return ManagedBuffer{llarp_buffer_t(payload)};
-        }
-      };
-
       bool
       HandleDataDrop(path::Path* p, const PathID_t& dst, uint64_t s);
 
@@ -216,54 +200,6 @@ namespace llarp
 
       bool
       ShouldBundleRC() const override;
-
-      struct SendContext
-      {
-        SendContext(const ServiceInfo& ident, const Introduction& intro,
-                    PathSet* send, Endpoint* ep);
-
-        void
-        AsyncEncryptAndSendTo(const llarp_buffer_t& payload, ProtocolType t);
-
-        /// send a fully encrypted hidden service frame
-        /// via a path on our pathset with path id p
-        bool
-        Send(const ProtocolFrame& f);
-
-        llarp::SharedSecret sharedKey;
-        ServiceInfo remoteIdent;
-        Introduction remoteIntro;
-        ConvoTag currentConvoTag;
-        PathSet* m_PathSet;
-        IDataHandler* m_DataHandler;
-        Endpoint* m_Endpoint;
-        uint64_t sequenceNo       = 0;
-        llarp_time_t lastGoodSend = 0;
-        llarp_time_t createdAt;
-        llarp_time_t sendTimeout    = 40 * 1000;
-        llarp_time_t connectTimeout = 60 * 1000;
-        bool markedBad              = false;
-
-        virtual bool
-        ShiftIntroduction(bool rebuild = true)
-        {
-          (void)rebuild;
-          return true;
-        };
-
-        virtual void
-        UpdateIntroSet(bool randomizePath = false) = 0;
-
-        virtual bool
-        MarkCurrentIntroBad(llarp_time_t now) = 0;
-
-       private:
-        void
-        EncryptAndSendTo(const llarp_buffer_t& payload, ProtocolType t);
-
-        virtual void
-        AsyncGenIntro(const llarp_buffer_t& payload, ProtocolType t) = 0;
-      };
 
       static void
       HandlePathDead(void*);
@@ -432,6 +368,9 @@ namespace llarp
       void
       PutNewOutboundContext(const IntroSet& introset);
 
+      uint64_t
+      GetSeqNoForConvo(const ConvoTag& tag);
+
       virtual void
       IntroSetPublishFail();
       virtual void
@@ -449,9 +388,6 @@ namespace llarp
 
       void
       PrefetchServicesByTag(const Tag& tag);
-
-      uint64_t
-      GetSeqNoForConvo(const ConvoTag& tag);
 
       bool
       IsolateNetwork();
